relay un-relayed packets on relayer start, also add command to do it manually
